django/requirements/base.txt
djangorestframework
markdown
django-filter
pymongo
pymemcache
boto3
psycopg2-binary
django-storages
django-tagging

